backend/requirements.txt
praw
gtts
moviepy==1.0.3
Flask
flask-cors
gunicorn
edge-tts
pydub
google-generativeai
python-dotenv